refactor: Config Property config 패키지로 이동
